dashboard

+++ system/library/cache/file.js
@@ -3,7 +3,7 @@ const path = require('path');
 function time() {
     return Math.floor(Date.now() / 1000);
 }
-module.exports = class File {
+module.exports = class CacheFileLibrary {
     constructor(expire = 3600) {
         this.expire = expire;
     }
@@ -26,22 +26,25 @@ module.exports = class File {
     delete(key) {
         const files = this.getMatchingFiles(key);
         for (const file of files) {
+            if(fs.existsSync(file))
             fs.unlinkSync(file);
         }
     }
     getFilePath(key, timestamp = null) {
-        return path.join(__dirname, 'cache', `cache.${key}.${timestamp ? timestamp : ''}`);
+        return path.join(DIR_CACHE, 'cache', `cache.${key}.${timestamp ? timestamp : ''}`);
     }
     getMatchingFiles(key) {
-        return fs.readdirSync(path.join(__dirname, 'cache')).filter(file => file.startsWith(`cache.${key}.`));
+        if (!fs.existsSync(path.join(DIR_CACHE, 'cache')))
+            fs.mkdirSync(path.join(DIR_CACHE, 'cache'), { recursive: true });
+        return fs.readdirSync(path.join(DIR_CACHE, 'cache')).filter(file => file.startsWith(`cache.${key}.`));
     }
     destructor() {
-        const files = fs.readdirSync(path.join(__dirname, 'cache'));
+        const files = fs.readdirSync(path.join(DIR_CACHE, 'cache'));
         if (Math.random() * 100 < 1) {
             for (const file of files) {
                 const timestamp = parseInt(file.split('.').pop());
                 if (timestamp < Date.now() / 1000) {
-                    fs.unlinkSync(path.join(__dirname, 'cache', file));
+                    fs.unlinkSync(path.join(DIR_CACHE, 'cache', file));
                 }
             }
         }
